power: smb5: Handle QC2.0 charger collapse gracefully
Adaptors that are non-compliant with QC2.0 are incapable of fulfilling
either a 9V or 12V request. In such cases, when the SMB5 makes such a
request, USB input collapse occurs - leading to an under-voltage
situation and suspension of the charger buck.
Fix this by by limiting the adaptor to a lower level, 5V or 9V
correspondingly. As an additional workaround, disable USBIN from
entering suspend upon collapse before re-running APSD, taking care to
re-enable it upon adaptor removal.

Change-Id: Iab9864b9c2d6e97296ad0e505c3384f29510e65f
Signed-off-by: Harry Yang <harryy@codeaurora.org>
Signed-off-by: Guru Das Srinagesh <gurus@codeaurora.org>
Signed-off-by: Ashish Chavan <ashichav@codeaurora.org>